% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openkmzzip.R
\name{openkmzzip}
\alias{openkmzzip}
\title{Open a zipped kmz format shapefile}
\usage{
openkmzzip(zipfile)
}
\arguments{
\item{file}{Name of zipped shapefile to be unpacked}
}
\value{
path of unzipped kml file
}
\description{
openkmzzip is created for the purpose of opening
shape files of Japanese census districts that are
published on  the Japanese government's e-Stat
website.
URL: https://www.e-stat.go.jp/gis/statmap-search?page=1&type=2&aggregateUnitForBoundary=A&toukeiCode=00200521

Requires: libexpat (https://libexpat.github.io/)
}
\details{

}
\examples{
shppath <- system.file("extdata",package="jpucd")
shpfile <- paste(shppath,list.files(shppath,pattern=".zip")[1],sep="/")
kmlfile <- openkmzzip(shpfile)

}
