% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOGRUTF8.R
\name{readOGRUTF8}
\alias{readOGRUTF8}
\title{Open a geojson format shapefile with UTF-8 encoding}
\usage{
readOGRUTF8(dsn, layer, encoding = "UTF-8", use_iconv = FALSE, ...)
}
\arguments{
\item{file}{Name of geojson shapefile to be unpacked}
}
\value{
Spatial object
}
\description{
readOGRUTF8 is created for the purpose of opening
shape files that have property data encoded in UTF-8
so that the property data can be read in RGUI
and written with writeOGR.

Background: use_iconv=TRUE in readOGR converts encoded
property fields to a data.frame that can be read on screen.
But when written using writeOGR, these fields are not
written. A workaround is to read the file with use_iconv=FALSE
then use Encoding(x) to set the encoding for data.frame columns.

Requires: rgdal (https://libexpat.github.io/)
}
\details{

}
\examples{

}
